Fix dos attack probability problem when liquidation
If there is too many gp and lp, the liquidation related function may
exceed the max gas limit in a transaction.

Don't transfer the liquidation token to gp or lp directly, let gp and
lp withdraw them.

In order to avoid some lps or gps  may forget this invested pool, but
this addresses must be very small, so we introduce a destroy interface
to return their token and reset the pool info.

Signed-off-by: Keefe-DeepGo <keefe.liu@deepquant.group>